Rename 'network' to 'models'
