+++ kglib/kgcn_experimental/models/embedding_test.py
@@ -23,7 +23,7 @@ import numpy as np
 import tensorflow as tf
 from mock import Mock
 from unittest.mock import patch
-from kglib.kgcn_experimental.network.embedding import common_embedding, attribute_embedding, node_embedding
+from kglib.kgcn_experimental.models.embedding import common_embedding, attribute_embedding, node_embedding
 from kglib.utils.test.utils import get_call_args
