+++ kglib/kgcn_experimental/models/typewise_IT.py
@@ -24,8 +24,8 @@ import tensorflow as tf
 from graph_nets.graphs import GraphsTuple
 from tensorflow.python.framework.ops import EagerTensor
 
-from kglib.kgcn_experimental.network.core import make_mlp_model
-from kglib.kgcn_experimental.network.typewise import TypewiseEncoder
+from kglib.kgcn_experimental.models.core import make_mlp_model
+from kglib.kgcn_experimental.models.typewise import TypewiseEncoder
 
 
 class ITTypewiseEncoder(unittest.TestCase):
