+++ kglib/kgcn_experimental/pipeline/pipeline.py
@@ -21,8 +21,8 @@ import networkx as nx
 import numpy as np
 from graph_nets.utils_np import graphs_tuple_to_networkxs
 
-from kglib.kgcn_experimental.network.attribute import CategoricalAttribute, BlankAttribute
-from kglib.kgcn_experimental.network.core import softmax, KGCN
+from kglib.kgcn_experimental.models.attribute import CategoricalAttribute, BlankAttribute
+from kglib.kgcn_experimental.models.core import softmax, KGCN
 from kglib.kgcn_experimental.learn.learn import KGCNLearner
 from kglib.kgcn_experimental.pipeline.utils import apply_logits_to_graphs, duplicate_edges_in_reverse
 from kglib.utils.graph.iterate import multidigraph_node_data_iterator, multidigraph_data_iterator
